requirements.txt
matplotlib==3.4.3
numpy==1.19.5
pandas==1.3.1
plotly==5.4.0
scikit_learn==1.1.3
setuptools==56.0.0
sklearn_pandas==2.2.0
